remove the explicit casting to (long) where possible, replacing constants by long ones
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@19341 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/node/MemoryChecker.java
@@ -47,7 +47,7 @@ public class MemoryChecker implements Runnable {
 			ps.queueTimedJob(this, 120 * 250); // 30 sec
 			return;
 		} else
-			ps.queueTimedJob(this, (long) 120 * sleeptime);
+			ps.queueTimedJob(this, 120L * sleeptime);
 		
 		// FIXME
 		// Do not remove until all known memory issues fixed,
